backend/pyproject.toml
[project]
name = "exam-transcripts-api"
version = "0.1.0"
description = "RESTful API for managing exam transcripts with role-based access control"
readme = "README.md"
authors = [
    { name = "Dieudonne Iyabivuze", email = "116635457+Iyabivuz-e@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=4.3.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501", "B008", "C408"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=app --cov-report=term-missing --cov-report=html"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.0",
]
